# mindmate-ai/code/api/src/routes/mood/index.ts
/**
 * Mood Routes
 * Mood tracking and analytics endpoints
 */

import { Router } from 'express';
import { validate } from '../../middleware/requestValidator';
import { authMiddleware } from '../../middleware/auth';
import { asyncHandler } from '../../middleware/errorHandler';
import {
  listMoodQuerySchema,
  logMoodSchema,
  updateMoodSchema,
  moodAnalyticsQuerySchema,
  moodIdParamSchema,
} from '../../validations';

const router = Router();

// Apply authentication middleware to all routes
router.use(authMiddleware);

// ==================== MOOD CRUD ====================

/**
 * @route   GET /mood
 * @desc    List mood entries
 * @access  Private
 */
router.get(
  '/',
  validate(listMoodQuerySchema, 'query'),
  asyncHandler(async (req, res) => {
    // TODO: Implement list mood entries
    res.json({
      entries: [],
      total: 0,
      average_score: 0,
      limit: 30,
      offset: 0,
      has_more: false,
    });
  })
);

/**
 * @route   POST /mood
 * @desc    Log mood
 * @access  Private
 */
router.post(
  '/',
  validate(logMoodSchema),
  asyncHandler(async (req, res) => {
    // TODO: Implement log mood
    res.status(201).json({
      id: null,
      user_id: null,
      mood_score: req.body.mood_score,
      mood_label: null,
      emotions: req.body.emotions || [],
      notes: req.body.notes || null,
      triggers: req.body.triggers || [],
      activities: req.body.activities || [],
      sleep_hours: req.body.sleep_hours || null,
      energy_level: req.body.energy_level || null,
      created_at: new Date().toISOString(),
      logged_date: new Date().toISOString().split('T')[0],
    });
  })
);

/**
 * @route   GET /mood/today
 * @desc    Get today's mood
 * @access  Private
 */
router.get(
  '/today',
  asyncHandler(async (req, res) => {
    // TODO: Implement get today's mood
    res.json({
      id: null,
      user_id: null,
      mood_score: null,
      mood_label: null,
      emotions: [],
      notes: null,
      triggers: [],
      activities: [],
      sleep_hours: null,
      energy_level: null,
      created_at: null,
      logged_date: new Date().toISOString().split('T')[0],
    });
  })
);

/**
 * @route   GET /mood/:entry_id
 * @desc    Get mood entry
 * @access  Private
 */
router.get(
  '/:entry_id',
  validate(moodIdParamSchema, 'params'),
  asyncHandler(async (req, res) => {
    // TODO: Implement get mood entry
    res.json({
      id: null,
      user_id: null,
      mood_score: null,
      mood_label: null,
      emotions: [],
      notes: null,
      triggers: [],
      activities: [],
      sleep_hours: null,
      energy_level: null,
      created_at: null,
      logged_date: null,
    });
  })
);

/**
 * @route   PATCH /mood/:entry_id
 * @desc    Update mood entry
 * @access  Private
 */
router.patch(
  '/:entry_id',
  validate(moodIdParamSchema, 'params'),
  validate(updateMoodSchema),
  asyncHandler(async (req, res) => {
    // TODO: Implement update mood entry
    res.json({
      id: null,
      user_id: null,
      mood_score: null,
      mood_label: null,
      emotions: [],
      notes: null,
      triggers: [],
      activities: [],
      sleep_hours: null,
      energy_level: null,
      created_at: null,
      logged_date: null,
    });
  })
);

/**
 * @route   DELETE /mood/:entry_id
 * @desc    Delete mood entry
 * @access  Private
 */
router.delete(
  '/:entry_id',
  validate(moodIdParamSchema, 'params'),
  asyncHandler(async (req, res) => {
    // TODO: Implement delete mood entry
    res.status(204).send();
  })
);

// ==================== MOOD ANALYTICS ====================

/**
 * @route   GET /mood/analytics
 * @desc    Get mood analytics
 * @access  Private
 */
router.get(
  '/analytics',
  validate(moodAnalyticsQuerySchema, 'query'),
  asyncHandler(async (req, res) => {
    // TODO: Implement mood analytics
    res.json({
      period: req.query.period || 'month',
      average_score: 0,
      highest_score: 0,
      lowest_score: 0,
      total_entries: 0,
      trend: 'stable',
      daily_averages: [],
      common_emotions: [],
      common_triggers: [],
    });
  })
);

/**
 * @route   GET /mood/patterns
 * @desc    Get mood patterns
 * @access  Private
 */
router.get(
  '/patterns',
  validate(moodAnalyticsQuerySchema, 'query'),
  asyncHandler(async (req, res) => {
    // TODO: Implement mood patterns
    res.json({
      patterns: [],
      insights: [],
    });
  })
);

/**
 * @route   GET /mood/streak
 * @desc    Get mood logging streak
 * @access  Private
 */
router.get(
  '/streak',
  asyncHandler(async (req, res) => {
    // TODO: Implement mood streak
    res.json({
      current_streak: 0,
      best_streak: 0,
      last_logged_date: null,
    });
  })
);

export default router;