# mindmate-ai/technical/api_spec.yaml
openapi: 3.0.3
info:
  title: MindMate AI API
  description: |
    # MindMate AI REST API
    
    Production-ready API specification for MindMate AI - An AI-powered mental health support platform.
    
    ## Features
    - **Authentication**: JWT-based authentication with refresh tokens
    - **AI Therapy Sessions**: Real-time conversational AI therapy
    - **Mood Tracking**: Daily mood logging and pattern analysis
    - **Task Management**: Personalized wellness tasks and goals
    - **Crisis Support**: Emergency resources and immediate assistance
    - **Progress Analytics**: Insights and streak tracking
    - **Notifications**: Smart reminders and wellness nudges
    
    ## Authentication
    All protected endpoints require a Bearer token in the Authorization header:
    ```
    Authorization: Bearer <jwt_token>
    ```
    
    ## Rate Limits
    - **Anonymous**: 10 requests/minute
    - **Authenticated**: 100 requests/minute
    - **Premium**: 500 requests/minute
    - **Admin**: 1000 requests/minute
    
    ## Error Codes
    | Code | Description |
    |------|-------------|
    | 400 | Bad Request - Invalid input |
    | 401 | Unauthorized - Invalid or missing token |
    | 403 | Forbidden - Insufficient permissions |
    | 404 | Not Found - Resource doesn't exist |
    | 409 | Conflict - Resource already exists |
    | 422 | Unprocessable Entity - Validation error |
    | 429 | Too Many Requests - Rate limit exceeded |
    | 500 | Internal Server Error |
  
...
servers:
...
paths:
...
  /mood:
    get:
      tags:
        - Mood
      summary: List mood entries
      description: Retrieve user's mood history
      operationId: listMoodEntries
      security:
        - bearerAuth: []
      parameters:
        - name: from_date
          in: query
          schema:
            type: string
            format: date
        - name: to_date
          in: query
          schema:
            type: string
            format: date
        - name: limit
          in: query
          schema:
            type: integer
            default: 30
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Mood entries retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MoodListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - Mood
      summary: Log mood
      description: Record a new mood entry
      operationId: logMood
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LogMoodRequest'
      responses:
        '201':
          description: Mood logged
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MoodEntry'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationError'
  /mood/today:
    get:
      tags:
        - Mood
      summary: Get today's mood
      description: Retrieve today's mood entry if exists
      operationId: getTodayMood
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Today's mood retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MoodEntry'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: No mood entry for today
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /mood/{entry_id}:
    get:
      tags:
        - Mood
      summary: Get mood entry
      description: Retrieve a specific mood entry
      operationId: getMoodEntry
      security:
        - bearerAuth: []
      parameters:
        - name: entry_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Mood entry retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MoodEntry'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags:
        - Mood
      summary: Update mood entry
      description: Update a mood entry (within 24 hours of creation)
      operationId: updateMoodEntry
      security:
        - bearerAuth: []
      parameters:
        - name: entry_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateMoodRequest'
      responses:
        '200':
          description: Mood entry updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MoodEntry'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Cannot update entry older than 24 hours
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags:
        - Mood
      summary: Delete mood entry
      description: Delete a mood entry
      operationId: deleteMoodEntry
      security:
        - bearerAuth: []
      parameters:
        - name: entry_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Mood entry deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /mood/analytics:
    get:
      tags:
        - Mood
      summary: Get mood analytics
      description: Retrieve mood patterns and insights
      operationId: getMoodAnalytics
      security:
        - bearerAuth: []
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [week, month, 3months, 6months, year]
            default: month
      responses:
        '200':
          description: Mood analytics retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MoodAnalytics'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /mood/patterns:
    get:
      tags:
        - Mood
      summary: Get mood patterns
      description: Retrieve AI-detected mood patterns and triggers
      operationId: getMoodPatterns
      security:
        - bearerAuth: []
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [week, month, 3months, 6months, year]
            default: month
      responses:
        '200':
          description: Mood patterns retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MoodPatterns'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /mood/streak:
    get:
      tags:
        - Mood
      summary: Get mood logging streak
      description: Retrieve current and best mood logging streaks
      operationId: getMoodStreak
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Mood streak retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  current_streak:
                    type: integer
                    description: Current consecutive days with mood entries
                  best_streak:
                    type: integer
                    description: Best streak ever achieved
                  last_logged_date:
                    type: string
                    format: date
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ==================== NOTIFICATIONS ====================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login/refresh endpoints
  responses:
...
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: unauthorized
            message: Invalid or missing authentication token
            code: 401
...
    NotFound:
      description: Not Found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: not_found
            message: Requested resource not found
            code: 404
...
    ValidationError:
      description: Validation Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationError'
...
  schemas:
...
    Error:
      type: object
      required:
        - error
        - message
        - code
      properties:
        error:
          type: string
          description: Error type identifier
        message:
          type: string
          description: Human-readable error message
        code:
          type: integer
          description: HTTP status code
        details:
          type: object
          description: Additional error details
        request_id:
          type: string
          description: Unique request ID for debugging
    ValidationError:
      allOf:
        - $ref: '#/components/schemas/Error'
        - type: object
          properties:
            field_errors:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  message:
                    type: string
                  code:
                    type: string

    # ==================== AUTH SCHEMAS ====================
...
    MoodEntry:
      type: object
      required:
        - id
        - user_id
        - mood_score
        - created_at
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        mood_score:
          type: integer
          minimum: 1
          maximum: 10
          description: 1 = Very low, 10 = Excellent
        mood_label:
          type: string
          enum: [terrible, bad, okay, good, great]
        emotions:
          type: array
          items:
            type: string
            enum: [happy, sad, anxious, angry, calm, excited, tired, stressed, grateful, frustrated, hopeful, lonely]
        notes:
          type: string
          maxLength: 2000
        triggers:
          type: array
          items:
            type: string
        activities:
          type: array
          items:
            type: string
            enum: [work, exercise, social, sleep, eating, meditation, hobby, family, outdoors, screen_time]
        sleep_hours:
          type: number
          minimum: 0
          maximum: 24
        energy_level:
          type: integer
          minimum: 1
          maximum: 10
        created_at:
          type: string
          format: date-time
        logged_date:
          type: string
          format: date
    LogMoodRequest:
      type: object
      required:
        - mood_score
      properties:
        mood_score:
          type: integer
          minimum: 1
          maximum: 10
        emotions:
          type: array
          items:
            type: string
        notes:
          type: string
          maxLength: 2000
        triggers:
          type: array
          items:
            type: string
        activities:
          type: array
          items:
            type: string
        sleep_hours:
          type: number
        energy_level:
          type: integer
          minimum: 1
          maximum: 10
    UpdateMoodRequest:
      type: object
      properties:
        mood_score:
          type: integer
          minimum: 1
          maximum: 10
        emotions:
          type: array
          items:
            type: string
        notes:
          type: string
          maxLength: 2000
        triggers:
          type: array
          items:
            type: string
        activities:
          type: array
          items:
            type: string
        sleep_hours:
          type: number
        energy_level:
          type: integer
    MoodListResponse:
      type: object
      required:
        - entries
        - total
      properties:
        entries:
          type: array
          items:
            $ref: '#/components/schemas/MoodEntry'
        total:
          type: integer
        average_score:
          type: number
        limit:
          type: integer
        offset:
          type: integer
        has_more:
          type: boolean
    MoodAnalytics:
      type: object
      properties:
        period:
          type: string
        average_score:
          type: number
        highest_score:
          type: integer
        lowest_score:
          type: integer
        total_entries:
          type: integer
        trend:
          type: string
          enum: [improving, declining, stable, fluctuating]
        daily_averages:
          type: array
          items:
            type: object
            properties:
              date:
                type: string
                format: date
              average_score:
                type: number
              entry_count:
                type: integer
        common_emotions:
          type: array
          items:
            type: object
            properties:
              emotion:
                type: string
              count:
                type: integer
              percentage:
                type: number
        common_triggers:
          type: array
          items:
            type: object
            properties:
              trigger:
                type: string
              count:
                type: integer
    MoodPatterns:
      type: object
      properties:
        patterns:
          type: array
          items:
            type: object
            properties:
              type:
                type: string
                enum: [time_of_day, day_of_week, activity_correlation, trigger_correlation]
              description:
                type: string
              confidence:
                type: number
              recommendation:
                type: string
        insights:
          type: array
          items:
            type: string

    # ==================== NOTIFICATION SCHEMAS ====================
...